logger: rate limit inside Docker
Also (these commits got combined at some point) avoid huge log messages on
logger copy failure

This is primarily a defense-in-depth measure, because we are also going to fix
journal.Send to not fail in this case.

Note that these days, Docker splits lines up before sending to our driver, so
this error is even less likely to occur. But let's try anyway.